Time: 229 ms (41.01%), Space: 74.2 MB (99.93%) - LeetHub

+++ 1894-find-the-student-that-will-replace-the-chalk/1894-find-the-student-that-will-replace-the-chalk.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int chalkReplacer(vector<int>& chalk,unsigned long long k) {
+    int chalkReplacer(vector<int>& chalk,int k) {
         long sum=0;
         for(auto &i:chalk)
             sum+=i;
